Refactor ETouchesService and use EventTableViewController for Home
+ Separate ETouchesService file into smaller files including:
Constants+eTouches, ETouchesAccessTOken
+ Have HomeNavigationController use EventTableViewController instead of
placeholder vc for dev purposes
+ Added missing comments on classes
+ Removed *.orig from .gitignore
