+++ .gitignore
@@ -1,5 +1,3 @@
 
 VOLA/VOLA/GoogleService-Info.plist
 
-
-*.orig
