+++ VOLA/VOLA/CalendarNavigationController.swift
@@ -2,6 +2,9 @@
 //  CalendarNavigationController.swift
 //  VOLA
 //
+//  CalendarNavigationController is the navigation controller for the calendar tab
+//  on the tab bar navigation. Here the root view controller is configured.
+//
 //  Created by Connie Nguyen on 6/16/17.
 //  Copyright © 2017 Systers-Opensource. All rights reserved.
 //
@@ -14,6 +17,7 @@ class CalendarNavigationController: UINavigationController {
         super.viewDidLoad()
 
         let eventTable = EventTableViewController.instantiateFromXib()
+        eventTable.tableType = .calendar
         setViewControllers([eventTable], animated: true)
     }
 }
