+++ VOLA/VOLA/Home.storyboard
@@ -12,15 +12,12 @@
         <!--Home-->
         <scene sceneID="mu1-Ev-8Ez">
             <objects>
-                <navigationController id="aYH-uM-SWg" sceneMemberID="viewController">
+                <navigationController id="aYH-uM-SWg" customClass="HomeNavigationController" customModule="VOLA" customModuleProvider="target" sceneMemberID="viewController">
                     <tabBarItem key="tabBarItem" title="Home" id="nzh-V4-Z8T"/>
-                    <navigationBar key="navigationBar" contentMode="scaleToFill" misplaced="YES" id="dZa-Yo-2iB">
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="dZa-Yo-2iB">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="44"/>
                         <autoresizingMask key="autoresizingMask"/>
                     </navigationBar>
-                    <connections>
-                        <segue destination="IWH-ZM-gGl" kind="relationship" relationship="rootViewController" id="n9A-lm-m6E"/>
-                    </connections>
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="QT9-kb-kDF" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
